api: rename checkUserHasRoleCertif => checkAdminMemberHasRoleCertif
