+++ api/lib/application/certification-courses/index.js
@@ -16,7 +16,7 @@ exports.register = async function (server) {
             method: (request, h) =>
               securityPreHandlers.userHasAtLeastOneAccessOf([
                 securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
                 securityPreHandlers.checkUserHasRoleSupport,
                 securityPreHandlers.checkUserHasRoleMetier,
               ])(request, h),
@@ -46,7 +46,7 @@ exports.register = async function (server) {
             method: (request, h) =>
               securityPreHandlers.userHasAtLeastOneAccessOf([
                 securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
                 securityPreHandlers.checkUserHasRoleSupport,
                 securityPreHandlers.checkUserHasRoleMetier,
               ])(request, h),
@@ -80,7 +80,7 @@ exports.register = async function (server) {
             method: (request, h) =>
               securityPreHandlers.userHasAtLeastOneAccessOf([
                 securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
                 securityPreHandlers.checkUserHasRoleSupport,
                 securityPreHandlers.checkUserHasRoleMetier,
               ])(request, h),
@@ -108,7 +108,7 @@ exports.register = async function (server) {
               securityPreHandlers.userHasAtLeastOneAccessOf([
                 securityPreHandlers.checkUserOwnsCertificationCourse,
                 securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
                 securityPreHandlers.checkUserHasRoleSupport,
               ])(request, h),
             assign: 'hasAuthorizationToAccessAdminScope',
@@ -139,7 +139,7 @@ exports.register = async function (server) {
               securityPreHandlers.userHasAtLeastOneAccessOf([
                 securityPreHandlers.checkUserOwnsCertificationCourse,
                 securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
                 securityPreHandlers.checkUserHasRoleSupport,
                 securityPreHandlers.checkUserHasRoleMetier,
               ])(request, h),
@@ -169,7 +169,7 @@ exports.register = async function (server) {
             method: (request, h) =>
               securityPreHandlers.userHasAtLeastOneAccessOf([
                 securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
                 securityPreHandlers.checkUserHasRoleSupport,
               ])(request, h),
             assign: 'hasAuthorizationToAccessAdminScope',
@@ -193,7 +193,7 @@ exports.register = async function (server) {
             method: (request, h) =>
               securityPreHandlers.userHasAtLeastOneAccessOf([
                 securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkUserHasRoleCertif,
+                securityPreHandlers.checkAdminMemberHasRoleCertif,
                 securityPreHandlers.checkUserHasRoleSupport,
               ])(request, h),
             assign: 'hasAuthorizationToAccessAdminScope',
