+++ api/tests/unit/application/certification-issue-reports/index_test.js
@@ -52,7 +52,7 @@ describe('Unit | Application | Certifications Issue Report | Route', function ()
         .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
         .withArgs([
           securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-          securityPreHandlers.checkUserHasRoleCertif,
+          securityPreHandlers.checkAdminMemberHasRoleCertif,
           securityPreHandlers.checkUserHasRoleSupport,
         ])
         .callsFake(
