+++ api/tests/unit/application/security-pre-handlers_test.js
@@ -3,7 +3,7 @@ const { expect, sinon, hFake, domainBuilder } = require('../../test-helper');
 const securityPreHandlers = require('../../../lib/application/security-pre-handlers');
 const tokenService = require('../../../lib/domain/services/token-service');
 const checkAdminMemberHasRoleSuperAdminUseCase = require('../../../lib/application/usecases/checkAdminMemberHasRoleSuperAdmin');
-const checkUserHasRoleCertifUseCase = require('../../../lib/application/usecases/checkUserHasRoleCertif');
+const checkAdminMemberHasRoleCertifUseCase = require('../../../lib/application/usecases/checkAdminMemberHasRoleCertif');
 const checkUserHasRoleSupportUseCase = require('../../../lib/application/usecases/checkUserHasRoleSupport');
 const checkUserHasRoleMetierUseCase = require('../../../lib/application/usecases/checkUserHasRoleMetier');
 const checkUserIsAdminInOrganizationUseCase = require('../../../lib/application/usecases/checkUserIsAdminInOrganization');
@@ -82,13 +82,13 @@ describe('Unit | Application | SecurityPreHandlers', function () {
     });
   });
 
-  describe('#checkUserHasRoleCertif', function () {
+  describe('#checkAdminMemberHasRoleCertif', function () {
     let hasRoleCertifStub;
     let request;
 
     beforeEach(function () {
       sinon.stub(tokenService, 'extractTokenFromAuthChain');
-      hasRoleCertifStub = sinon.stub(checkUserHasRoleCertifUseCase, 'execute');
+      hasRoleCertifStub = sinon.stub(checkAdminMemberHasRoleCertifUseCase, 'execute');
       request = { auth: { credentials: { accessToken: 'valid.access.token', userId: 1234 } } };
     });
 
@@ -101,7 +101,7 @@ describe('Unit | Application | SecurityPreHandlers', function () {
         // given
 
         // when
-        const response = await securityPreHandlers.checkUserHasRoleCertif(request, hFake);
+        const response = await securityPreHandlers.checkAdminMemberHasRoleCertif(request, hFake);
 
         // then
         expect(response.source).to.equal(true);
@@ -114,7 +114,7 @@ describe('Unit | Application | SecurityPreHandlers', function () {
         delete request.auth.credentials;
 
         // when
-        const response = await securityPreHandlers.checkUserHasRoleCertif(request, hFake);
+        const response = await securityPreHandlers.checkAdminMemberHasRoleCertif(request, hFake);
 
         // then
         expect(response.statusCode).to.equal(403);
@@ -123,10 +123,10 @@ describe('Unit | Application | SecurityPreHandlers', function () {
 
       it('should forbid resource access when user does not have role Certif', async function () {
         // given
-        checkUserHasRoleCertifUseCase.execute.resolves(false);
+        checkAdminMemberHasRoleCertifUseCase.execute.resolves(false);
 
         // when
-        const response = await securityPreHandlers.checkUserHasRoleCertif(request, hFake);
+        const response = await securityPreHandlers.checkAdminMemberHasRoleCertif(request, hFake);
 
         // then
         expect(response.statusCode).to.equal(403);
@@ -135,10 +135,10 @@ describe('Unit | Application | SecurityPreHandlers', function () {
 
       it('should forbid resource access when an error is thrown by use case', async function () {
         // given
-        checkUserHasRoleCertifUseCase.execute.rejects(new Error('Some error'));
+        checkAdminMemberHasRoleCertifUseCase.execute.rejects(new Error('Some error'));
 
         // when
-        const response = await securityPreHandlers.checkUserHasRoleCertif(request, hFake);
+        const response = await securityPreHandlers.checkAdminMemberHasRoleCertif(request, hFake);
 
         // then
         expect(response.statusCode).to.equal(403);
