+++ api/tests/unit/application/session/index_test.js
@@ -541,7 +541,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
@@ -590,7 +590,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
@@ -641,7 +641,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
@@ -730,7 +730,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
@@ -789,7 +789,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
@@ -910,7 +910,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
@@ -953,7 +953,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
@@ -981,7 +981,7 @@ describe('Unit | Application | Sessions | Routes', function () {
           .stub(securityPreHandlers, 'userHasAtLeastOneAccessOf')
           .withArgs([
             securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkUserHasRoleCertif,
+            securityPreHandlers.checkAdminMemberHasRoleCertif,
             securityPreHandlers.checkUserHasRoleSupport,
           ])
           .callsFake(
